# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  email: string;
}

export interface AuthResponse {
  user: AdminUser;
  message: string;
}

export type ShortLinkStatus = typeof ShortLinkStatus[keyof typeof ShortLinkStatus];


export const ShortLinkStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ShortLink {
  id: number;
  code: string;
  name: string;
  destinationUrl: string;
  status: ShortLinkStatus;
  /** @nullable */
  expiryDate: string | null;
  clicks: number;
  scans: number;
  totalEngagements: number;
  createdAt: string;
  updatedAt: string;
}

export type ShortLinkInputStatus = typeof ShortLinkInputStatus[keyof typeof ShortLinkInputStatus];


export const ShortLinkInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ShortLinkInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  destinationUrl: string;
  status: ShortLinkInputStatus;
  /** @nullable */
  expiryDate?: string | null;
}

export type ShortLinkUpdateStatus = typeof ShortLinkUpdateStatus[keyof typeof ShortLinkUpdateStatus];


export const ShortLinkUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ShortLinkUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  destinationUrl?: string;
  status?: ShortLinkUpdateStatus;
  /** @nullable */
  expiryDate?: string | null;
}

export interface QrCode {
  dataUrl: string;
  shortUrl: string;
  code: string;
}

export interface AnalyticsSummary {
  totalEngagements: number;
  totalActiveCodes: number;
  totalInactiveCodes: number;
  expiredCodes: number;
  totalLinks: number;
}

export type LinkAnalyticsStatus = typeof LinkAnalyticsStatus[keyof typeof LinkAnalyticsStatus];


export const LinkAnalyticsStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface LinkAnalytics {
  id: number;
  code: string;
  name: string;
  clicks: number;
  scans: number;
  total: number;
  status: LinkAnalyticsStatus;
}

export type ListLinksParams = {
/**
 * Filter by status
 */
status?: ListLinksStatus;
/**
 * Search by name or code
 */
search?: string;
/**
 * Filter by expiry
 */
expiry?: ListLinksExpiry;
};

export type ListLinksStatus = typeof ListLinksStatus[keyof typeof ListLinksStatus];


export const ListLinksStatus = {
  active: 'active',
  inactive: 'inactive',
  all: 'all',
} as const;

export type ListLinksExpiry = typeof ListLinksExpiry[keyof typeof ListLinksExpiry];


export const ListLinksExpiry = {
  all: 'all',
  expired: 'expired',
  notExpired: 'notExpired',
} as const;