# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalEngagements: number;
  totalActiveCodes: number;
  totalInactiveCodes: number;
  expiredCodes: number;
  totalLinks: number;
}