# lib/api-zod/src/generated/types/linkAnalytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { LinkAnalyticsStatus } from './linkAnalyticsStatus';

export interface LinkAnalytics {
  id: number;
  code: string;
  name: string;
  clicks: number;
  scans: number;
  total: number;
  status: LinkAnalyticsStatus;
}