# lib/api-zod/src/generated/types/shortLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ShortLinkStatus } from './shortLinkStatus';

export interface ShortLink {
  id: number;
  code: string;
  name: string;
  destinationUrl: string;
  status: ShortLinkStatus;
  /** @nullable */
  expiryDate: string | null;
  clicks: number;
  scans: number;
  totalEngagements: number;
  createdAt: string;
  updatedAt: string;
}