# lib/api-zod/src/generated/types/shortLinkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ShortLinkInputStatus } from './shortLinkInputStatus';

export interface ShortLinkInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  destinationUrl: string;
  status: ShortLinkInputStatus;
  /** @nullable */
  expiryDate?: string | null;
}