# lib/api-zod/src/generated/types/shortLinkUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QR Link Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ShortLinkUpdateStatus } from './shortLinkUpdateStatus';

export interface ShortLinkUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  destinationUrl?: string;
  status?: ShortLinkUpdateStatus;
  /** @nullable */
  expiryDate?: string | null;
}